+++ src/main.rs
@@ -1,12 +1,11 @@
 mod member;
 mod opts;
 mod v3;
+mod v4;
 
 use crate::member::{Member, MemberBuilder};
 use crate::opts::{Opt, SpgFileVersion};
-use async_std::io::prelude::*;
-use async_std::io::{self, Read, Result};
-use oxidized_mdf::MdfDatabase;
+use async_std::prelude::*;
 use structopt::StructOpt;
 
 #[async_std::main]
@@ -17,15 +16,12 @@ async fn main() -> std::io::Result<()> {
 
     let mut members = match opt.file_version {
         SpgFileVersion::V3 => v3::parse(read).await?,
-        SpgFileVersion::V4 => {
-            // parse_v4(read).await?;
-            todo!()
-        }
+        SpgFileVersion::V4 => v4::parse(read).await?,
     };
 
     members.sort();
 
-    let mut out = io::stdout();
+    let mut out = async_std::io::stdout();
     out.write_all(b"[").await?;
     for (i, m) in members.iter().enumerate() {
         if i > 0 {
@@ -35,17 +31,6 @@ async fn main() -> std::io::Result<()> {
         out.write_all(json.as_bytes()).await?;
     }
     out.write_all(b"]").await?;
-
-    Ok(())
-}
-
-async fn parse_v4(read: Box<dyn Read + Unpin>) -> Result<()> {
-    let database = MdfDatabase::from_read(read).await?;
-
-    let mut stdout = io::stdout();
-    stdout
-        .write_all(database.database_name().as_bytes())
-        .await?;
 
     Ok(())
 }